Problem: index discovery mechanism is hard-coded
It couples the discovery with definition and is not well suited for providing
different mechanisms to load indices, particularly other JVM languages.

Solution: extract an IndexLoader interface and move current implementation to
JavaStaticFieldIndexLoader.
