+++ eventsourcing-core/src/main/java/com/eventsourcing/index/IndexLoader.java
@@ -1,0 +1,14 @@
+/**
+ * Copyright (c) 2016, All Contributors (see CONTRIBUTORS file)
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.eventsourcing.index;
+
+import com.googlecode.cqengine.index.Index;
+
+public interface IndexLoader {
+    Iterable<Index> load(IndexEngine engine, Class klass) throws IndexEngine.IndexNotSupported;
+}
